changed directory name to cdk to avoid Cargo.toml errors
